frontend/pyproject.toml
[build-system]
requires = ["setuptools>=45", "setuptools_scm>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "disaster-prediction"
version = "0.1.0"
description = "Disaster Risk Prediction Application"
requires-python = ">=3.8"
dependencies = [
    "streamlit",
    "torch",
    "numpy",
    "scikit-learn",
    "pandas",
    "pyyaml",
    "plotly",
    "geopandas"
]

[tool.setuptools]
packages = ["frontend", "frontend.streamlit_app", "frontend.streamlit_app.components"]